writing-practice/requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
Pillow==10.2.0
python-dotenv==1.0.1
flake8==7.0.0
black==24.1.1
isort==5.13.2
pytest==8.0.2
pytest-mock==3.12.0
